Make the code more modular
Add checksums
