+++ dev-tools/release/utils/checksums.py
@@ -1,0 +1,40 @@
+# Licensed to Elasticsearch under one or more contributor
+# license agreements. See the NOTICE file distributed with
+# this work for additional information regarding copyright
+# ownership. Elasticsearch licenses this file to you under
+# the Apache License, Version 2.0 (the "License"); you may
+# not use this file except in compliance  with the License.
+# You may obtain a copy of the License at
+#
+# http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing,
+# software distributed under the License is distributed on
+# an 'AS IS' BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND,
+# either express or implied. See the License for the specific
+# language governing permissions and limitations under the License.
+
+# Reads the given file and applies the
+# callback to it. If the callback changed
+# a line the given file is replaced with
+# the modified input.
+
+import os
+
+
+# Generates sha1 for a file
+# and returns the checksum files as well
+# as the given files in a list
+def generate_checksums(release_file):
+    res = []
+    directory = os.path.dirname(release_file)
+    file = os.path.basename(release_file)
+    checksum_file = '%s.sha1.txt' % file
+
+    if os.system('cd %s; shasum %s > %s' % (directory, file, checksum_file)):
+        raise RuntimeError('Failed to generate checksum for file %s' % release_file)
+    res += [os.path.join(directory, checksum_file), release_file]
+    return res
+
+
+
